change type of RoomCode from uint8_t to uint16_t

+++ shared/shared/model/RoomCode.h
@@ -2,5 +2,5 @@
 #include <cstdint>
 
 namespace card {
-	typedef std::uint8_t RoomCode;
+	typedef std::uint16_t RoomCode;
 }
